requirements.txt
asgiref==3.6.0
Django==4.1.7
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
mysqlclient==2.1.1
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
